Rename: FavoriteButton 폴더 이동
